LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-shows-start-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MacOSInlineMediaControls</code> showsStartButton property.");

const mediaControls = new MacOSInlineMediaControls;

debug("Default where showsStartButton = false");
shouldBeFalse("mediaControls.showsStartButton");
shouldBe("mediaControls.children", "[mediaControls.topLeftControlsBar, mediaControls.bottomControlsBar]");

debug("");
debug("showsStartButton = true");
mediaControls.showsStartButton = true;
shouldBeTrue("mediaControls.showsStartButton");
shouldBe("mediaControls.children", "[mediaControls.playPauseButton]");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
