Safari-611.1.10/LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-shows-start-button-expected.txt

Testing the MacOSInlineMediaControls showsStartButton property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Default where showsStartButton = false
PASS mediaControls.showsStartButton is false
PASS mediaControls.children is [mediaControls.topLeftControlsBar, mediaControls.bottomControlsBar]

showsStartButton = true
PASS mediaControls.showsStartButton is true
PASS mediaControls.children is [mediaControls.playPauseButton]

PASS successfullyParsed is true

TEST COMPLETE
